+++ spec/atom-environment-spec.js
@@ -1,4 +1,4 @@
-const { conditionPromise } = require('./async-spec-helpers');
+const { conditionPromise } = require('./helpers/async-spec-helpers');
 const fs = require('fs');
 const path = require('path');
 const temp = require('temp').track();
